chore: move legacy resolver layer
